First iteration create profil leaving form
This is working, but need some more code to be fully coherent.

When leaving a page, extract from form a profil.
If this profil does not already exist, create it in front.

As unload event kill process quickly, don't have time to
create it in back.

So in front, a temporary profil is saved, which will
be created later in back with event "load".