2 major fix
Fix bug when having an empty field to create new profil.
This was causing wront profil db content, wrong call API
and breaking profil creation on load.

Fix bug to check profil exitence.
In the case of missing key for a profil to compare, was breaking code.2 major fix

Fix bug when having an empty field to create new profil.
This was causing wront profil db content, wrong call API
and breaking profil creation on load.

Fix bug to check profil exitence.
In the case of missing key for a profil to compare, was breaking code.